# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Certif",
        "description": "Api de l'appli PENC",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authentifier l'utilisateur et obtenir le jeton JWT",
                "description": "Authentifier un utilisateur en fournissant son adresse e-mail et son mot de passe, et obtenir un jeton JWT en réponse",
                "operationId": "44212a9096e4b09358281e9ec8a0701d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "utilisateur@exemple.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie, jeton JWT reçu"
                    },
                    "401": {
                        "description": "Non autorisé, identifiants invalides"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Enregistrer un nouvel utilisateur",
                "description": "Enregistrer un nouvel utilisateur en fournissant les informations requises",
                "operationId": "a718f172ff3ac464d723835815f8fb57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nouveau@utilisateur.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscription réussie, utilisateur créé"
                    },
                    "422": {
                        "description": "Échec de la validation des données"
                    }
                }
            }
        },
...
        "/api/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Rafraîchir le jeton JWT actuel",
                "description": "Rafraîchir le jeton JWT actuel en obtenant un nouveau jeton valide",
                "operationId": "7d5da2870d3eba95cdd67878ab20da3a",
                "responses": {
                    "200": {
                        "description": "Rafraîchissement réussi, nouveau jeton JWT reçu"
                    },
                    "401": {
                        "description": "Non autorisé, jeton JWT invalide"
                    }
                }
            }
        },
...
        "/api/addfield": {
            "post": {
                "tags": [
                    "Fields"
                ],
                "summary": "Stocke une nouvelle ressource créée",
                "description": "Stocke une nouvelle ressource créée dans le stockage (Il sera utilisé pour stocker dans notre base de données les champs nouvellement créés).",
                "operationId": "3db3f7101991eda4f4aac366d2923e9d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFieldRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Opération réussie"
                    }
                }
            }
        },
        "/api/deletefield/{field}": {
            "delete": {
                "tags": [
                    "Fields"
                ],
                "summary": "Supprime la ressource spécifiée",
                "description": "Supprime la ressource spécifiée.",
                "operationId": "e6b82a28e722f51a8539b72a5a6da5c9",
                "parameters": [
                    {
                        "name": "field",
                        "in": "path",
                        "description": "ID de la ressource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Field"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/addforum": {
            "post": {
                "tags": [
                    "Forums"
                ],
                "summary": "Stocke une nouvelle ressource créée",
                "description": "Stocke une nouvelle ressource créée dans le stockage (Il sera utilisé pour stocker dans notre base de données les forums nouvellement créés).",
                "operationId": "fa3b0667189c49d549a2137bf0bc69f5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreForumRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Votre forum a bien été créé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entité non traitable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Il existe déjà un forum pour ce domaine"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/deleteforum/{forum}": {
            "delete": {
                "tags": [
                    "Forums"
                ],
                "summary": "Supprime la ressource spécifiée",
                "description": "Supprime la ressource spécifiée.",
                "operationId": "d9e02131e2b5d7cc5466056a8dd9e9cb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID de la ressource à supprimer",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le forum a été supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/displaymessage": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Retourne la liste des sujets avec leurs messages associés",
                "description": "Retourne la liste des sujets avec leurs messages associés.",
                "operationId": "6d879defb5f48e74a1e221a53bf3ced9",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sendmessage": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Crée un nouveau message dans un sujet spécifié",
                "description": "Crée un nouveau message dans un sujet spécifié.",
                "operationId": "514defc4d6d1297e61e693997583ab54",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_content",
                                    "topic_id"
                                ],
                                "properties": {
                                    "message_content": {
                                        "type": "string",
                                        "example": "Contenu du message"
                                    },
                                    "topic_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Votre message a bien été créé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Un administrateur ne peut pas envoyer de message."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Une erreur est survenue lors de la création du message."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/updatespecificmessage/{message}": {
            "put": {
                "tags": [
                    "Messages"
                ],
                "summary": "Met à jour un message existant",
                "description": "Met à jour un message existant.",
                "operationId": "1d8f083ca53736d8103878f424c9dfbf",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "ID du message à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_content",
                                    "user_id",
                                    "topic_id"
                                ],
                                "properties": {
                                    "message_content": {
                                        "type": "string",
                                        "example": "Nouveau contenu du message"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "topic_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Modifications effectuées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Modifications effectuées avec succès."
                                        },
                                        "resultat": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ce message ne vous appartient pas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne. Veuillez réessayer."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Message d'erreur spécifique"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/displayreply": {
            "get": {
                "tags": [
                    "Replies"
                ],
                "summary": "Récupère tous les messages et leurs réponses",
                "description": "Récupère tous les messages et leurs réponses.",
                "operationId": "b06c7812fc8b25d84806a5e7ae0879d5",
                "responses": {
                    "200": {
                        "description": "Liste de messages et leurs réponses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Reply"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Une erreur inattendue est survenue"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}